Use Cassette to track _all_ func calls, to regenerate StagedFunctions when _any_ dependent calls are invalidated
We trace the execution used to compute the generated function body, to
get exactly all the functions that need backedges to the generator.

This is the most naive and simplest implementation to correctly get
StagedFunctions that are recompiled when any dependent function is
invalidated! :)

Future optimizations could include:
- Not using Cassette, since it will cause recompilation of every
function call in the entire call tree.
- Only adding edges to functions that don't already get edges added by
inference. (I think that this basically maps 1:1 with dynamic
dispatches -- so if we could somehow instrument the dynamic dispatches
triggered during execution of the generated function body, we could use
just those functions as the edges.)
-